+++ CHANGELOG.rst
@@ -14,6 +14,7 @@ FORMS Feature
 * :+1: Added nice to have `Forms` methods such as `has_part` and `compatible_within_scope` and tests (#1108)
 * :+1: Removed the `target_scope` parameter when cloning inside the same scope
 * :+1: Extended the `DashboardWidget` with the form numbers and charts (#1123)
+* :+1: Added the concept of a sidebar card to the sidebar.
 
 v3.18.0 (22MAR22)
 -----------------